backend/requirements.txt
Django==4.2.21
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
redis==5.0.1
ffmpeg-python==0.2.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4 